Revert "blackbox: update blackbox to 0.77"
This reverts commit a285722d376b0d7fc41c63c1ac35899fe395ca6b.

Autoconf 2.71 makes too big troubles with gtk packages as it i.e. madates
full build/install of gtk-doc.
MiniMyth2 uses minimal gtk packages. To avoid full build of gtk
packages - lets go back to autoconf 2.69.
This will force downgrade of packages which can't be build with
autoconf 2.69. One of them is this package - so revert it to
version building with autoconf 2.69.